Extracted `ClusterProfileWidget` component from `ClusterProfilesWidget`:   * To handle `Cluster Info` panel expansion state.
